ui tweaks and refactoring
