+++ plugins/cloud-carbon-footprint/src/components/CloudCarbonFootprintPlugin.test.tsx
@@ -14,13 +14,14 @@
  * limitations under the License.
  */
 import React from 'react';
-import { render } from '@testing-library/react';
-import { ExampleFetchComponent } from './ExampleFetchComponent';
+import { CloudCarbonFootprintPlugin } from './CloudCarbonFootprintPlugin';
+import { ThemeProvider } from '@material-ui/core';
+import { lightTheme } from 'packages/theme';
 import { rest } from 'msw';
 import { setupServer } from 'msw/node';
-import { setupRequestMockHandlers } from '@backstage/test-utils';
+import { setupRequestMockHandlers, renderInTestApp } from 'packages/test-utils';
 
-describe('ExampleFetchComponent', () => {
+describe('ExampleComponent', () => {
   const server = setupServer();
   // Enable sane handlers for network requests
   setupRequestMockHandlers(server);
@@ -28,13 +29,16 @@ describe('ExampleFetchComponent', () => {
   // setup mock response
   beforeEach(() => {
     server.use(
-      rest.get('https://randomuser.me/*', (_, res, ctx) =>
-        res(ctx.status(200), ctx.delay(2000), ctx.json({})),
-      ),
+      rest.get('/*', (_, res, ctx) => res(ctx.status(200), ctx.json({}))),
     );
   });
+
   it('should render', async () => {
-    const rendered = render(<ExampleFetchComponent />);
-    expect(await rendered.findByTestId('progress')).toBeInTheDocument();
+    const rendered = await renderInTestApp(
+      <ThemeProvider theme={lightTheme}>
+        <CloudCarbonFootprintPlugin />
+      </ThemeProvider>,
+    );
+    expect(rendered.getByText('Cloud Carbon Footprint')).toBeInTheDocument();
   });
 });
